+++ src/ipts/control.hpp
@@ -20,7 +20,7 @@ public:
 
 	void send_feedback();
 	u32 doorbell();
-	int read(void *buf, size_t size);
+	ssize_t read(void *buf, size_t size);
 	void reset();
 
 private:
